Speed: Double =>Float

+++ client/src/main/scala/com/neo/sk/gypsy/model/GridOnClient.scala
@@ -295,7 +295,7 @@ class GridOnClient(override val boundary: Point) extends Grid {
             hasMoved =true
             newMass += p.mass
             newRadius = Mass2Radius(newMass)
-            newSpeed = sqrt(pow(vx,2)+ pow(vy,2))
+            newSpeed = sqrt(pow(vx,2)+ pow(vy,2)).toFloat
             newTargetX = vx.toShort
             newTargetY = vy.toShort
             massList = massList.filterNot(l => l == p)
